Moved definitions to new file, removed simulator specific definitions
